+++ docroot/js/simplicity/simplicityGoogleMapBoundsTracker.js
@@ -134,9 +134,9 @@
         radiusMeters2 = google.maps.geometry.spherical.computeDistanceBetween(
             center, new google.maps.LatLng(bounds.getNorthEast().lat(), center.lng()));
       } else {
-        radiusMeters1 = $.simplicityGeoFn.distanceKm(
+        radiusMeters1 = $.simplicity.haversineDistanceKm(
             center.lat(), center.lng(), center.lat(), bounds.getNorthEast().lng()) * 1000.0;
-        radiusMeters2 = $.simplicityGeoFn.distanceKm(
+        radiusMeters2 = $.simplicity.haversineDistanceKm(
             center.lat(), center.lng(), bounds.getNorthEast().lat(), center.lng()) * 1000.0;
       }
       var minMeters = Math.min(radiusMeters1, radiusMeters2);